+++ packages/util/src/array/chunk.spec.ts
@@ -1,0 +1,12 @@
+// Copyright 2017-2020 @polkadot/util authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+import { arrayChunk } from '.';
+
+describe('arrayChunk', (): void => {
+  it('chunks arrays', (): void => {
+    expect(
+      arrayChunk([1, 2, 3, 4, 5, 6, 7], 3)
+    ).toEqual([[1, 2, 3], [4, 5, 6], [7]]);
+  });
+});