+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Changes:
 
 - Add `arrayFlatten` to merge a series of arrays into a single
+- Add `arrayChunk` to chunk an array into multiples of defined size
 
 
 ## 5.1.1 Dec 19, 2020